+++ src/main/java/domain/player/Player.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 public record Player(
         String name,
